+++ elmyra.ip.access.epo/CHANGES.rst
@@ -10,6 +10,7 @@ development
 - ui: temporarily remove cql quick query builder helper actions
 - ui: make pagination links black, not blue
 - ui: fix link to CCD Viewer (upgrade from /CCD-2.0.0 to /CCD-2.0.4)
+- ui: print/pdf: honor current query and pagesize
 
 0.6.1
 =====